ovn-tester/ovn_workload: Move Namespace class to CMS plugin.
