+++ ovn-tester/cms/ovn_kubernetes/ovn_kubernetes.py
@@ -10,16 +10,335 @@ import ovn_utils
 import ovn_stats
 
 from ovn_utils import DualStackSubnet
-from ovn_workload import (
-    ChassisNode,
-    Cluster,
-    DEFAULT_BACKEND_PORT,
-    DEFAULT_VIP_PORT,
-)
+from ovn_workload import ChassisNode, Cluster
 
 log = logging.getLogger(__name__)
 ClusterBringupCfg = namedtuple('ClusterBringupCfg', ['n_pods_per_node'])
 OVN_HEATER_CMS_PLUGIN = 'OVNKubernetesCluster'
+ACL_DEFAULT_DENY_PRIO = 1
+ACL_DEFAULT_ALLOW_ARP_PRIO = 2
+ACL_NETPOL_ALLOW_PRIO = 3
+DEFAULT_NS_VIP_SUBNET = netaddr.IPNetwork('30.0.0.0/16')
+DEFAULT_NS_VIP_SUBNET6 = netaddr.IPNetwork('30::/32')
+DEFAULT_VIP_PORT = 80
+DEFAULT_BACKEND_PORT = 8080
+
+
+class Namespace:
+    def __init__(self, clusters, name, global_cfg):
+        self.clusters = clusters
+        self.nbctl = [cluster.nbctl for cluster in clusters]
+        self.ports = [[] for _ in range(len(clusters))]
+        self.enforcing = False
+        self.pg_def_deny_igr = [
+            nbctl.port_group_create(f'pg_deny_igr_{name}')
+            for nbctl in self.nbctl
+        ]
+        self.pg_def_deny_egr = [
+            nbctl.port_group_create(f'pg_deny_egr_{name}')
+            for nbctl in self.nbctl
+        ]
+        self.pg = [
+            nbctl.port_group_create(f'pg_{name}') for nbctl in self.nbctl
+        ]
+        self.addr_set4 = [
+            (
+                nbctl.address_set_create(f'as_{name}')
+                if global_cfg.run_ipv4
+                else None
+            )
+            for nbctl in self.nbctl
+        ]
+        self.addr_set6 = [
+            (
+                nbctl.address_set_create(f'as6_{name}')
+                if global_cfg.run_ipv6
+                else None
+            )
+            for nbctl in self.nbctl
+        ]
+        self.sub_as = [[] for _ in range(len(clusters))]
+        self.sub_pg = [[] for _ in range(len(clusters))]
+        self.load_balancer = None
+        for cluster in self.clusters:
+            cluster.n_ns += 1
+        self.name = name
+
+    @ovn_stats.timeit
+    def add_ports(self, ports, az=0):
+        self.ports[az].extend(ports)
+        # Always add port IPs to the address set but not to the PGs.
+        # Simulate what OpenShift does, which is: create the port groups
+        # when the first network policy is applied.
+        if self.addr_set4:
+            for i, nbctl in enumerate(self.nbctl):
+                nbctl.address_set_add_addrs(
+                    self.addr_set4[i], [str(p.ip) for p in ports]
+                )
+        if self.addr_set6:
+            for i, nbctl in enumerate(self.nbctl):
+                nbctl.address_set_add_addrs(
+                    self.addr_set6[i], [str(p.ip6) for p in ports]
+                )
+        if self.enforcing:
+            self.nbctl[az].port_group_add_ports(
+                self.pg_def_deny_igr[az], ports
+            )
+            self.nbctl[az].port_group_add_ports(
+                self.pg_def_deny_egr[az], ports
+            )
+            self.nbctl[az].port_group_add_ports(self.pg[az], ports)
+
+    def unprovision(self):
+        # ACLs are garbage collected by OVSDB as soon as all the records
+        # referencing them are removed.
+        for i, cluster in enumerate(self.clusters):
+            cluster.unprovision_ports(self.ports[i])
+        for i, nbctl in enumerate(self.nbctl):
+            nbctl.port_group_del(self.pg_def_deny_igr[i])
+            nbctl.port_group_del(self.pg_def_deny_egr[i])
+            nbctl.port_group_del(self.pg[i])
+            if self.addr_set4:
+                nbctl.address_set_del(self.addr_set4[i])
+            if self.addr_set6:
+                nbctl.address_set_del(self.addr_set6[i])
+            nbctl.port_group_del(self.sub_pg[i])
+            nbctl.address_set_del(self.sub_as[i])
+
+    def unprovision_ports(self, ports, az=0):
+        '''Unprovision a subset of ports in the namespace without having to
+        unprovision the entire namespace or any of its network policies.'''
+
+        for port in ports:
+            self.ports[az].remove(port)
+
+        self.clusters[az].unprovision_ports(ports)
+
+    def enforce(self):
+        if self.enforcing:
+            return
+        self.enforcing = True
+        for i, nbctl in enumerate(self.nbctl):
+            nbctl.port_group_add_ports(self.pg_def_deny_igr[i], self.ports[i])
+            nbctl.port_group_add_ports(self.pg_def_deny_egr[i], self.ports[i])
+            nbctl.port_group_add_ports(self.pg[i], self.ports[i])
+
+    def create_sub_ns(self, ports, global_cfg, az=0):
+        n_sub_pgs = len(self.sub_pg[az])
+        suffix = f'{self.name}_{n_sub_pgs}'
+        pg = self.nbctl[az].port_group_create(f'sub_pg_{suffix}')
+        self.nbctl[az].port_group_add_ports(pg, ports)
+        self.sub_pg[az].append(pg)
+        for i, nbctl in enumerate(self.nbctl):
+            if global_cfg.run_ipv4:
+                addr_set = nbctl.address_set_create(f'sub_as_{suffix}')
+                nbctl.address_set_add_addrs(
+                    addr_set, [str(p.ip) for p in ports]
+                )
+                self.sub_as[i].append(addr_set)
+            if global_cfg.run_ipv6:
+                addr_set = nbctl.address_set_create(f'sub_as_{suffix}6')
+                nbctl.address_set_add_addrs(
+                    addr_set, [str(p.ip6) for p in ports]
+                )
+                self.sub_as[i].append(addr_set)
+        return n_sub_pgs
+
+    @ovn_stats.timeit
+    def default_deny(self, family, az=0):
+        self.enforce()
+
+        addr_set = f'self.addr_set{family}.name'
+        self.nbctl[az].acl_add(
+            self.pg_def_deny_igr[az].name,
+            'to-lport',
+            ACL_DEFAULT_DENY_PRIO,
+            'port-group',
+            f'ip4.src == \\${addr_set} && '
+            f'outport == @{self.pg_def_deny_igr[az].name}',
+            'drop',
+        )
+        self.nbctl[az].acl_add(
+            self.pg_def_deny_egr[az].name,
+            'to-lport',
+            ACL_DEFAULT_DENY_PRIO,
+            'port-group',
+            f'ip4.dst == \\${addr_set} && '
+            f'inport == @{self.pg_def_deny_egr[az].name}',
+            'drop',
+        )
+        self.nbctl[az].acl_add(
+            self.pg_def_deny_igr[az].name,
+            'to-lport',
+            ACL_DEFAULT_ALLOW_ARP_PRIO,
+            'port-group',
+            f'outport == @{self.pg_def_deny_igr[az].name} && arp',
+            'allow',
+        )
+        self.nbctl[az].acl_add(
+            self.pg_def_deny_egr[az].name,
+            'to-lport',
+            ACL_DEFAULT_ALLOW_ARP_PRIO,
+            'port-group',
+            f'inport == @{self.pg_def_deny_egr[az].name} && arp',
+            'allow',
+        )
+
+    @ovn_stats.timeit
+    def allow_within_namespace(self, family, az=0):
+        self.enforce()
+
+        addr_set = f'self.addr_set{family}.name'
+        self.nbctl[az].acl_add(
+            self.pg[az].name,
+            'to-lport',
+            ACL_NETPOL_ALLOW_PRIO,
+            'port-group',
+            f'ip4.src == \\${addr_set} && ' f'outport == @{self.pg[az].name}',
+            'allow-related',
+        )
+        self.nbctl[az].acl_add(
+            self.pg[az].name,
+            'to-lport',
+            ACL_NETPOL_ALLOW_PRIO,
+            'port-group',
+            f'ip4.dst == \\${addr_set} && ' f'inport == @{self.pg[az].name}',
+            'allow-related',
+        )
+
+    @ovn_stats.timeit
+    def allow_cross_namespace(self, ns, family):
+        self.enforce()
+
+        for az, nbctl in enumerate(self.nbctl):
+            if len(self.ports[az]) == 0:
+                continue
+            addr_set = f'self.addr_set{family}.name'
+            nbctl[az].acl_add(
+                self.pg[az].name,
+                'to-lport',
+                ACL_NETPOL_ALLOW_PRIO,
+                'port-group',
+                f'ip4.src == \\${addr_set} && '
+                f'outport == @{ns.pg[az].name}',
+                'allow-related',
+            )
+            ns_addr_set = f'ns.addr_set{family}.name'
+            nbctl[az].acl_add(
+                self.pg[az].name,
+                'to-lport',
+                ACL_NETPOL_ALLOW_PRIO,
+                'port-group',
+                f'ip4.dst == \\${ns_addr_set} && '
+                f'inport == @{self.pg[az].name}',
+                'allow-related',
+            )
+
+    @ovn_stats.timeit
+    def allow_sub_namespace(self, src, dst, family, az=0):
+        self.nbctl[az].acl_add(
+            self.pg[az].name,
+            'to-lport',
+            ACL_NETPOL_ALLOW_PRIO,
+            'port-group',
+            f'ip{family}.src == \\${self.sub_as[az][src].name} && '
+            f'outport == @{self.sub_pg[az][dst].name}',
+            'allow-related',
+        )
+        self.nbctl[az].acl_add(
+            self.pg[az].name,
+            'to-lport',
+            ACL_NETPOL_ALLOW_PRIO,
+            'port-group',
+            f'ip{family}.dst == \\${self.sub_as[az][dst].name} && '
+            f'inport == @{self.sub_pg[az][src].name}',
+            'allow-related',
+        )
+
+    @ovn_stats.timeit
+    def allow_from_external(
+        self, external_ips, include_ext_gw=False, family=4, az=0
+    ):
+        self.enforce()
+        # If requested, include the ext-gw of the first port in the namespace
+        # so we can check that this rule is enforced.
+        if include_ext_gw:
+            assert len(self.ports) > 0
+            if family == 4 and self.ports[az][0].ext_gw:
+                external_ips.append(self.ports[az][0].ext_gw)
+            elif family == 6 and self.ports[az][0].ext_gw6:
+                external_ips.append(self.ports[az][0].ext_gw6)
+        ips = [str(ip) for ip in external_ips]
+        self.nbctl[az].acl_add(
+            self.pg[az].name,
+            'to-lport',
+            ACL_NETPOL_ALLOW_PRIO,
+            'port-group',
+            f'ip.{family} == {{{",".join(ips)}}} && '
+            f'outport == @{self.pg[az].name}',
+            'allow-related',
+        )
+
+    @ovn_stats.timeit
+    def check_enforcing_internal(self, az=0):
+        # "Random" check that first pod can reach last pod in the namespace.
+        if len(self.ports[az]) > 1:
+            src = self.ports[az][0]
+            dst = self.ports[az][-1]
+            worker = src.metadata
+            if src.ip:
+                worker.ping_port(self.clusters[az], src, dst.ip)
+            if src.ip6:
+                worker.ping_port(self.clusters[az], src, dst.ip6)
+
+    @ovn_stats.timeit
+    def check_enforcing_external(self, az=0):
+        if len(self.ports[az]) > 0:
+            dst = self.ports[az][0]
+            worker = dst.metadata
+            worker.ping_external(self.clusters[az], dst)
+
+    @ovn_stats.timeit
+    def check_enforcing_cross_ns(self, ns, az=0):
+        if len(self.ports[az]) > 0 and len(ns.ports[az]) > 0:
+            dst = ns.ports[az][0]
+            src = self.ports[az][0]
+            worker = src.metadata
+            if src.ip and dst.ip:
+                worker.ping_port(self.clusters[az], src, dst.ip)
+            if src.ip6 and dst.ip6:
+                worker.ping_port(self.clusters[az], src, dst.ip6)
+
+    def create_load_balancer(self, az=0):
+        self.load_balancer = lb.OvnLoadBalancer(
+            f'lb_{self.name}', self.nbctl[az]
+        )
+
+    @ovn_stats.timeit
+    def provision_vips_to_load_balancers(self, backend_lists, version, az=0):
+        vip_ns_subnet = DEFAULT_NS_VIP_SUBNET
+        if version == 6:
+            vip_ns_subnet = DEFAULT_NS_VIP_SUBNET6
+        vip_net = vip_ns_subnet.next(self.clusters[az].n_ns)
+        n_vips = len(self.load_balancer.vips.keys())
+        vip_ip = vip_net.ip.__add__(n_vips + 1)
+
+        if version == 6:
+            vips = {
+                f'[{vip_ip + i}]:{DEFAULT_VIP_PORT}': [
+                    f'[{p.ip6}]:{DEFAULT_BACKEND_PORT}' for p in ports
+                ]
+                for i, ports in enumerate(backend_lists)
+            }
+            self.load_balancer.add_vips(vips)
+        else:
+            vips = {
+                f'{vip_ip + i}:{DEFAULT_VIP_PORT}': [
+                    f'{p.ip}:{DEFAULT_BACKEND_PORT}' for p in ports
+                ]
+                for i, ports in enumerate(backend_lists)
+            }
+            self.load_balancer.add_vips(vips)
 
 
 class OVNKubernetesCluster(Cluster):
